Improve export UI
