+++ frontend/src/app/pages/exercises/exercise/shared/time-travel/time-travel.component.html
@@ -73,6 +73,12 @@
                 >
                     Statistiken
                 </button>
+                <button
+                    (click)="createNewExerciseFromTheCurrentState()"
+                    class="btn btn-primary btn-sm ms-5"
+                >
+                    Neue Übung an dieser Stelle erstellen
+                </button>
             </div>
         </div>
         <div class="pt-1">
