+++ frontend/src/app/pages/exercises/exercise/shared/time-travel/time-travel.component.ts
@@ -1,8 +1,13 @@
 import type { OnDestroy } from '@angular/core';
 import { Component } from '@angular/core';
 import { NgbModal } from '@ng-bootstrap/ng-bootstrap';
+import { Store } from '@ngrx/store';
+import { StateExport } from 'digital-fuesim-manv-shared';
 import { throttle } from 'lodash';
 import { ApiService } from 'src/app/core/api.service';
+import { MessageService } from 'src/app/core/messages/message.service';
+import type { AppState } from 'src/app/state/app.state';
+import { getStateSnapshot } from 'src/app/state/get-state-snapshot';
 import { openClientOverviewModal } from '../client-overview/open-client-overview-modal';
 import { openExerciseStatisticsModal } from '../exercise-statistics/open-exercise-statistics-modal';
 import { openTransferOverviewModal } from '../transfer-overview/open-transfer-overview-modal';
@@ -15,7 +20,9 @@ import { openTransferOverviewModal } from '../transfer-overview/open-transfer-ov
 export class TimeTravelComponent implements OnDestroy {
     constructor(
         private readonly modalService: NgbModal,
-        public readonly apiService: ApiService
+        public readonly apiService: ApiService,
+        private readonly store: Store<AppState>,
+        private readonly messageService: MessageService
     ) {}
 
     public openClientOverview() {
@@ -88,6 +95,21 @@ export class TimeTravelComponent implements OnDestroy {
         this.stopReplay();
         this.apiService.jumpToTime(this.apiService.timeConstraints!.start);
         this.startReplay();
+    }
+
+    public async createNewExerciseFromTheCurrentState() {
+        const currentState = getStateSnapshot(this.store).exercise;
+        const { trainerId } = await this.apiService.importExercise(
+            new StateExport(currentState)
+        );
+        this.messageService.postMessage({
+            color: 'success',
+            title: 'Neue Übung erstellt',
+            body: `ÜbungsleiterId: ${trainerId}`,
+        });
+        window
+            .open(`${location.origin}/exercises/${trainerId}`, '_blank')
+            ?.focus();
     }
 
     ngOnDestroy() {
